leadforge/requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
duckduckgo-search>=6.2.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
pydantic>=2.7.0
